refactor: OpenFiles UI를 File 도메인으로 이동
