+++ src/features/File/OpenFiles/ui/FileNavPanel.tsx
