+++ src/features/File/OpenFiles/ui/IDEScrollView.tsx
